version 2.12_p1

+++ RootReader/Reader_dd/reader_v2.cpp
@@ -58,7 +58,7 @@ public:
   void analyse(Nuball2Tree & tree, Event & event);
   void write();
 
-  static void setMaxHits(int const & nb_max_hits) {g_nb_max_hits = nb_max_hits;}
+  static void setMaxHits(long const & nb_max_hits) {g_nb_max_hits = nb_max_hits;}
 
 private:
   /// @brief static
